Creating alert DTOs
