+++ patricare-back/src/modules/alert/dto/create-alert.dto.ts
@@ -1,0 +1,15 @@
+import { IsString, IsOptional } from 'class-validator';
+
+export class CreateAlerteDto {
+  @IsString() titre: string;
+  @IsString() commentaire: string;
+
+  @IsOptional()
+  userId?: number;
+
+  @IsOptional()
+  logementId?: number;
+
+  @IsOptional()
+  typeId?: number;
+}
